chore: refactor local formatter in SelectorGrid to common component
