+++ apps/studio/components/interfaces/TableGridEditor/SidePanelEditor/RowEditor/ForeignRowSelector/SelectorGrid.tsx
@@ -2,8 +2,7 @@ import { Key } from 'lucide-react'
 import DataGrid, { Column } from 'react-data-grid'
 import { Tooltip, TooltipContent, TooltipTrigger } from 'ui'
 
-import { convertByteaToHex } from '../RowEditor.utils'
-import { NullValue } from '@/components/grid/components/common/NullValue'
+import { SelectGridCellFormatter } from '@/components/grid/components/formatter/SelectGridCellFormatter'
 import { COLUMN_MIN_WIDTH } from '@/components/grid/constants'
 import type { SupaRow } from '@/components/grid/types'
 import {
@@ -36,28 +35,6 @@ const columnRender = (name: string, isPrimaryKey = false) => {
   )
 }
 
-// TODO: move this formatter out to a common component
-const formatter = ({ column, format, row }: { column: string; format: string; row: SupaRow }) => {
-  const formattedValue =
-    format === 'bytea'
-      ? convertByteaToHex(row[column])
-      : row[column] === null
-        ? null
-        : typeof row[column] === 'object'
-          ? JSON.stringify(row[column])
-          : row[column]
-
-  return (
-    <div className="group sb-grid-select-cell__formatter overflow-hidden">
-      {formattedValue === null ? (
-        <NullValue />
-      ) : (
-        <span className="text-sm truncate">{formattedValue}</span>
-      )}
-    </div>
-  )
-}
-
 const SelectorGrid = ({ rows, onRowSelect }: SelectorGridProps) => {
   const snap = useTableEditorTableStateSnapshot()
 
@@ -71,8 +48,9 @@ const SelectorGrid = ({ rows, onRowSelect }: SelectorGridProps) => {
     const result: Column<SupaRow> = {
       key: column.name,
       name: column.name,
-      renderCell: (props) =>
-        formatter({ column: column.name, format: column.format, row: props.row }),
+      renderCell: (props) => (
+        <SelectGridCellFormatter column={column.name} format={column.format} row={props.row} />
+      ),
       renderHeaderCell: () => columnRender(column.name, column.isPrimaryKey),
       resizable: true,
       width: columnWidth,
